// src/components/Image/Image.d.ts
export interface ImageProps {
    src: string;
    alt: string;
    width?: string;
    height?: string;
    rounded?: boolean;
    shadow?: boolean;
    border?: boolean;
    className?: string;
}
declare function Image({ src, alt, width, height, rounded, shadow, border, className, }: ImageProps): import("react/jsx-runtime").JSX.Element;
export default Image;
//# sourceMappingURL=Image.d.ts.map